+++ CHANGELOG.rst
@@ -13,3 +13,4 @@ Changelog
 * Added realtime search.
 * Removed logging.
 * Added command to show available trending locations.
+* Added command to see what's trending.